requirements.txt
astroid==1.6.0
Django==2.0.1
django-crispy-forms==1.7.0
django-filter==1.1.0
djangorestframework==3.7.7
isort==4.2.15
lazy-object-proxy==1.3.1
mccabe==0.6.1
mock==2.0.0
pbr==3.1.1
psycopg2==2.7.4
pylint==1.8.1
pylint-django==0.8.0
pylint-plugin-utils==0.2.6
pytz==2017.3
six==1.11.0
sqlparse==0.2.2
urlfetch==1.0.2
wrapt==1.10.11
